properly send rooms and application through link
