+++ frontend/src/components/AdminApplicationOverview/index.tsx
@@ -4,16 +4,20 @@ import { AnyAction } from 'redux';
 import { ThunkAction, ThunkDispatch } from 'redux-thunk';
 import { IStore } from '../../store';
 import { IApplication } from '../Application';
+import { IRoom } from '../ViewRooms';
+import { fetchAllRooms } from '../ViewRooms/actions';
 import { fetchAllApplications } from './actions';
 import './adminApplicationOverview.css';
 import Presentational from './Presentational';
 
 interface IStateProps {
   applications: IApplication[];
+  rooms: IRoom[];
 }
 
 interface IDispatchProps {
   getAllApplications: () => ThunkAction<void, {}, {}, AnyAction>;
+  getRooms: () => ThunkAction<void, {}, {}, AnyAction>;
 }
 
 type Props = IStateProps & IDispatchProps;
@@ -22,22 +26,26 @@ type Props = IStateProps & IDispatchProps;
 class _Container extends React.Component<Props> {
   public componentDidMount() {
     this.props.getAllApplications();
+    this.props.getRooms();
   }
 
   public render() {
     return (
       <Presentational
         applications={this.props.applications}
+        rooms={this.props.rooms}
       />);
   }
 }
 
 const mapStateToProps = (state: IStore) => ({
   applications: state.applications.applications,
+  rooms: state.rooms.rooms,
 });
 
 const mapDispatchToProps = (dispatch: ThunkDispatch<{}, {}, any>) => ({
   getAllApplications: () => dispatch(fetchAllApplications()),
+  getRooms: () => dispatch(fetchAllRooms()),
 });
 
 const Container = connect(
